some minor code cleanup
